chore: improve yarn lock error checking (#17001)
The yarn lock checking script was always wrong: it would assume a
package was local if its version number was `0.0.0`... but this is
no longer true after running `align-version.sh`.

I don't understand why this doesn't fail on the build server, but
it definitely fails for me on my machine when trying to do a full
build "as if" I was the build server.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*