docs(core): fix CfnMapping example (#16882)
The CloudFormation intrinsic function `Fn::FindInMap` only supports alphanumeric characters as name. However, the `CfnMapping` examples in the README file contain hyphens in the name field. This causes an error when the code is deployed.

I changed the structure of the examples to create examples that can be deployed in AWS.

Fixes #16866.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*