chore: include RELEASE_NOTES in packed distribution
We now create and generate a `RELEASE_NOTES.md` file as part of our build, with
the intent to use it for our GitHub release publishing. See (#16942) and
https://github.com/cdklabs/aws-delivlib/pull/1044.

What I missed is that after the build, we pack, and then only include what ends
up in `./dist` in the final build artifact that goes to publishing. This fix
includes the release notes in the dist folder so it will be picked up and used
by the release process.