fix(rds): using both Instance imports & exports for Postgres fails deployment (#17060)
Our CDK code was assuming that all instance engines that support S3 imports & exports
must use the same Role if both functions are enabled at the same time.
However, it turns out that's true only for Oracle and SQL Server,
but not for Postgres - in fact, Postgres has the exact opposite requirement,
and will fail deployment if the same Role is used for both.
Change our code to only use a single Role if the engine is Oracle or SQL Server.

Fixes #16757

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*