feat(servicecatalog): Add `ProductStack` resource and ability to create a product version from cdk code.
Add ability to define a product version entirely within CDK as opposed to referencing templates or local assets.
The service catalog `ProductStack` is similar to `NestedStacks` that do not deploy themselves but rather are referenced
by the parent stacks.  The resources defined in your product are added to the product stack like any other cdk app.

Co-authored-by: Dillon Ponzo <dponzo18@gmail.com>