chore: fix master build by skipping optional release notes copy (#17154)
For releases (v1 and v2), we need to create the release notes (as part of the
build) and then copy them to the dist/ directory (as part of pack). However, for
`master` builds, we don't create the release notes because the BUMP_CANDIDATE
builds reference a non-existent release version. Assuming the presence of the
release notes in pack is causing the `master` pipeline to fail.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*