chore: update DEPRECATED_APIs.md and introduce new plaintext format (#17120)
As part of the upcoming v2 GA, we need to update (and lock in) the final list of
deprecated APIs that will be removed from V2. At the same time, we need to
create the input for `jsii`'s new `--strip-deprecated` file input, where only a
subset of deprecated items are stripped away (see #16566). This change does both
by introducing a new output format for `list-deprecated-apis` and feeding that
value into `jsii` at compile time.

fixes #16566


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*