chore(CDK V2 canaries): allow canary tests to run with a specific aws-cdk version (#17285)
If an environment variable, `$RELEASE_TAG` is defined, it will be appended to the end of the `npm install aws-cdk@` command. This will allow us to run canaries that test against v2 by specifying `RELEASE_TAG=next` in the environment variables. If `$RELEASE_TAG` is not defined, `npm install aws-cdk@latest` will be used. 

This PR also includes a comment in `release-notes.ts` that the v2 publishing verification canary is dependent on the format of the release notes. 
 
part of #16593, see https://github.com/cdklabs/cdk-ops/pull/1769 

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*