fix(cli): no longer disable rollback by default for hotswap deployments
Previously, we would default the --rollback flag to `false` if it wasn't provided
if the `--hotswap` option was also passed to `cdk deploy`.
However, disabled rollback does not play well with deployments that cause replacements -
they fail, and leave the Stack in a weird state.
For that reason, stop disabling rollback by default.
It's still possible to disable easily if wanted by using the `-R` switch.

Fixes #17267