feat(ecs): ecs services can be created with an imported task definition
ECS Services can now be created with imported Task Definitions. The ECS
Service constructs make use of information contained in the task
definition in order to make it easier for the user to configure their
ECS Service. When using an imported task definition, this information is
not available. This PR addresses that limitation by requiring the user
to provide that information when it is required. For example, when
adding the ECS Service as a target of a Load Balancer, the user will be
required to provide the `containerName`, `containerPort`, and `hostPort`
since that is information that is contained in the ContainerDefinition,
and not available on an imported TaskDefinition.

As part of this change the `addToTaskRolePolicy` method was moved to the
base class so it is now available to be used with imported
TaskDefinitions.

There are two alternatives to this approach. The first is to simply
throw an error any time the user attempts an action that requires this
missing information. The second is to add a new optional property to the
`fromXXXTaskDefinitionAttributes` method for the `defaultContainer` that
contains the required information. This second method would be more
involved and would most likely also require the creation of an
`IContainerDefinition` interface.

closes #7863,#6240