chore: fix packages transform for stability stripping (#17388)
In https://github.com/aws/aws-cdk/pull/17327, we stripped away the stability banner for `Cfn` resources from README's of alpha modules, since these modules don't actually contain L1 resources. 

Problem is that this stripping causes a bunch of `pkglint` violations on the `v2-main` branch, for example:

```console
@aws-cdk/aws-iot-alpha: In package package.json
--
972 | @aws-cdk/aws-iot-alpha: - [package-info/maturity] Missing stability banner for experimental in README.md file (fixable)
973 | @aws-cdk/aws-iot-alpha: Error: Some package.json files had errors
974 | @aws-cdk/aws-iot-alpha:     at main (/codebuild/output/src115445882/src/github.com/aws/aws-cdk/tools/@aws-cdk/pkglint/bin/pkglint.js:30:15)
975 | @aws-cdk/aws-iot-alpha:     at Object.<anonym
```

This PR fixes the transform and makes the necessary changes so that `pkglint` will be happy. 
Note that this was not detected during merge of https://github.com/aws/aws-cdk/pull/17327 into master since on master, these packages are transformed into private packages, and hence are not validated. 

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*