feat(lambda-nodejs): custom asset hash (#16412)
By default the asset hash for a `NodejsFunction` is based on the
bundling output, the result of the `esbuild` bundling.

Add a `assetHash` prop to customize the asset hash and allow users to
implement their own logic. This can be useful if users want to invalidate
the asset based on specific (input) file contents or logic.


Closes #16157


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*