chore: new workflow that adds a checklist when there is a change to custom resources (#17302)
For PRs that contain changes to custom resources, we want to draw the author's attention to the part of the documentation that explains the important cases to handle. This workflow will create a checklist with a single item, pointing to that section of the README, in case the PR contains keywords such as `CustomResource` or `CustomResourceProvider`.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*