chore(ubergen): make rosetta fixtures submodule aware (#17438)
Currently ubergen bundles all fixtures into the same `rosetta` folder.
This means that all fixtures in files called `default.ts-fixture` replace
each other as they are found and vie to be the last `default.ts-fixture`
standing in the `rosetta` folder. This causes unintended compiliation
failures for `yarn rosetta:extract --compile` in projects that rely on
ubergen, since they will be copied into the wrong `default.ts-fixture`.

The solution here in jsii is to find fixtures based on submodules: https://github.com/aws/jsii/pull/3131.
This PR updates ubergen to construct the necessary subfolders in
`rosetta` so that the expected submodule fixtures can be found. 

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*