chore(cli): correctly set up `cdk watch` in init templates (#17440)
This enables `watch` to work out-of-the-box by adding the `watch` key to all init template `cdk.json` files. By default `watch` will look only at source files; project configuration files will not be included by `watch`. The `build` key is not needed for any language, as all of the `app` commands will build the code before it is executed. 

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*