chore(redshift): fix build break due to undeclared dependency
The final installation tests and documentation generation is failing due to:
`Error: Cannot find module 'aws-sdk/clients/redshiftdata'`

This is because the SDK is an implicit/assumed dependency for Lambda and being
used in a custom resource. However, some of the construct library code took a
dependency on the Lambda code that included this library, triggering the
dependency to (attempt to) be loaded.

This is a fundamental issue with the design of this feature -- there is code
shared between the construct library and custom resource (in both directions),
which causes some fragility we're not detecting in our normal build. While the
broader picture deserves some thought, this specific example can be tactically
dealt with to fix the build and unblock our CI/CD pipeline.