chore(lambda-go): remove all go code from integ test handler (#17504)
The tests of `aws-lambda-go` use packages that depend on `gopkg.in`,
a webservice that allows some form of indirection of package names
and versions to GitHub commits.

Unfortunately, this website is having trouble, failing our builds with
errors like:

```
@aws-cdk/aws-lambda-go: go: gopkg.in/check.v1@v1.0.0-20200902074654-038fdea0a05b: invalid version: git fetch -f origin refs/heads/*:refs/heads/* refs/tags/*:refs/tags/* in /go/pkg/mod/cache/vcs/9241c28341fcedca6a799ab7a465dd6924dc5d94044cbfabb75778817250adfc: exit status 128:
@aws-cdk/aws-lambda-go:     error: RPC failed; HTTP 502 curl 22 The requested URL returned error: 502 Bad Gateway
```

This website doesn't look like it's being run by a large organization
with engineering resources. Rather, it looks like a one-person operation
who cannot guarantee uptime. This threatens our builds for no benefit.

Eradicate all traces of `gopkg.in` in the go dependency set. Since `github.com/aws/aws-lamda-go`
depends on `gopkg.in/yaml.v3`, we have to get rid of that dependency, and therefore get rid of
all Go handler code. It doesn't need to run anyway, it just needs to build.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*