fix(aws-lambda-event-sources): `Function.addEventSource` fails for `ManagedKafkaEventSource` typed parameters (#17490)
----
This request is to fix the issue with creating an event source mapping for a lambda function at the same time of MSK cluster creation. The existing code requires the MSK cluster to be created first and the replacing the clusterArn with a hard coded value. This is due to the use of the clusterArn in the construct Id which is still a token value at time of synth. Replaced the clusterArn in the Id with just the topic and the lambda function name that is associated to the mapping. This follows similar functionality as the Kinesis event source and I continued the pattern for testing to expose the eventMappingId to ensure it is defined when the event source is mapped to the lambda function. fixes #15700

Thanks @otaviomacedo for pushing me to make the change.

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*