fix(cli): improve asset publishing times by up to 30% (#17409)
Part of #17266

Improves the performance of asset publishing by caching the AWS SDK clients that are used, based on which roles need to be assumed. This way, most roles only need to be assumed once per `cdk deploy`, instead once per asset.

In https://github.com/cdklabs/construct-hub, we have a CDK stack that includes 20+ assets (at least 15 of which are Lambda code bundles). I tested out the updated CLI, and it sped up the publishing step (when no assets needed to be changed) from 29.8 seconds to 19.6 seconds.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*