chore(ubergen): examples in v2 libraries have incorrect imports (#17524)
Ubergen rewrites imports for rosetta fixtures using `rewriteImports()`, which used to return a relative path to the file. This would fail compilation when running `rosetta:extract --compile` on v2. This PR adds a callback function "`newImport()`" to `rewriteImports()` that actually does the work of creating a new import statement. This allows the caller of `rewriteImports()` to determine the path to the file, whether that is a relative one or an external one. 

Rosetta fixtures have imports like:

```ts
import * as appscaling from "monocdk/aws-applicationautoscaling"; // Automatically re-written from '@aws-cdk/aws-applicationautoscaling'
```

and

```ts
import * as lambda from "aws-cdk-lib/aws-lambda"; // Automatically re-written from '@aws-cdk/aws-lambda'
```
----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*