fix(cloudwatch): render agnostic alarms in legacy style
Render legacy style alarms even if the user explicitly passes in the
`Aws.ACCOUNT_ID` token: we can compare these strings since we every call
to `Aws.ACCOUNT_ID` is guaranteed to return the same tokenized string.
Only switch to modern-style alarms if we cannot successfully determine
that they are definitely equal.

This simplifies one more case that would accidentally trigger new-style
alarms, which some SNS notification consumers are not equipped to handle.

Add a section to the README explaining the differences between old and
new-style alarms, and that SNS consumers need to pay attention to the
differences.