chore: ignore asset hashes in integ tests by default
Any random change that changes asset hashes is a pain to deal with,
and slows us down massively. We have a pragma to ignore hashes,
and we just turn that pragma on in most integ tests.

Might as well flip the default around then: ignore the hashes UNLESS
requested otherwise.

Do the hashes help? Theoretically during our latest upgrade of jsii,
the upgrade that incorrectly added extra statements to our Lambda's
sources that *happen* to also be compiled with jsii, we could have
used the hashes to detect this undesirable change. But did we? No,
we just updated the hashes and went on with our life, because there
are so many hashes to update all the time that we don't think twice
about it. Therefore these hashes in shapshots are not a useful
mechanism, they only slow us down for no gain.