feat(servicecatalog): Add tag options to a CloudformationProduct
Users can now associate tag options to a cloudformation product through an association call
or upon instantiation. Tag options addded to a portfolio are made available for any products within it,
but you can also have separate, product level tag options.  We only create unique tag option constructs in the template
but we can have the same tag option associated with both a portfolio and a product in that portfolio, the logic that
resolves this is handled by service catalog

Co-authored-by: Dillon Ponzo <dponzo18@gmail.com>