chore(rds): undeprecated `SnapshotCredentials.fromGeneratedPassword()`
All deprecated APIs will be removed from CDKv2.

The `SnapshotCredentials.fromGeneratedPassword()` will modify the RDS
instance in ways that may impact usability of the resource. This API
must not be deprecated.

The alternative APIs to the `DatabaseInstanceEngine` APIs refereced in
this PR will cause the [CFN EngineVersion][1] to be modified.
Modification of this property causes [some interruption][2] to the
resource.
This may cause "some interruption" to users' running applications.

[1]: https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-rds-database-instance.html#cfn-rds-dbinstance-engineversion
[2]: https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/using-cfn-updating-stacks-update-behaviors.html#update-some-interrupt