chore(rds): undeprecated APIs whose migration will cause interruption (#17683)
All deprecated APIs will be removed from CDKv2.

Migrating from `SnapshotCredentials.fromGeneratedPassword()` to
its documented alternative will modify the RDS
instance in ways that may impact usability of the resource. This API
must not be deprecated.

The alternative APIs to the `DatabaseInstanceEngine` APIs refereced in
this PR will cause the [CFN EngineVersion][1] to be modified.
Modification of this property causes [some interruption][2] to the
resource.
This may cause "some interruption" to users' running applications.

[1]: https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-rds-database-instance.html#cfn-rds-dbinstance-engineversion
[2]: https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/using-cfn-updating-stacks-update-behaviors.html#update-some-interrupt


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*