fix(codepipeline): cannot trigger on all tags anymore in EcrSourceAction (#17270)
The EcrSourceAction could previously be used to trigger on changes to all tags of an image. As part of the fix aws#13818, the imageTag was defaulted to latest if not provided. Therefore it was no longer possible to call the underlying onCloudTrailImagePushed function with an undefined imageTag to watch changes on all tags.

Reintroduce triggering on all tags by passing an empty string as the imageTag.

Fixes aws#13818


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*