chore: fix list of deprecated APIs so APIs are actually stripped in V2.
The change to strip deprecated APIs based on a list was introduced several weeks
ago; however, that list was built off of the v1/master branch, not v2. This
meant the FQNs of all of the types was wrong, leading to *nothing* being
stripped, and all deprecated APIs being re-introduced into the V2 source.

This change re-builds the deprecated_apis.txt list so the correct APIs are
removed from V2.