chore(assertions): snippets to avoid using unneeded variables
Previously, the rosetta translations could not determine the type of
objects being passed into methods when they are untyped. Some were being
interpreted as 'props' type while they should just be regarding as
`Record` or `any`.

To compensate for this, the README in this module assigned them to
variables, so the translator did a better job at knowing this.

This has now been fixed in rosetta. Move back to using this inline,
since the usage is just nicer.